requirements.txt
requests
beautifulsoup4
psycopg2-binary
curl-cffi>=0.7
